Moves AirtableImporter into own subdirectory (client)
